ocr-python/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.1.0
requests==2.32.3
beautifulsoup4==4.12.3
python-multipart==0.0.20
psycopg2-binary==2.9.10
